refactor(theme): migrate code related to theme under the theme package
