--- packages/app/src/components/index.ts
@@ -1,3 +1,0 @@
-export { default as ThemeButton } from './ThemeButton';
-export { default as ThemeIcon } from './ThemeIcon';
-export { default as ThemeSwitch } from './ThemeSwitch';
