--- packages/app/src/contexts/index.ts
@@ -1,1 +1,0 @@
-export * from './theme';
