+++ packages/theme/src/index.ts
@@ -1,0 +1,4 @@
+export * from './theme';
+export { default as ThemeButton } from './ThemeButton';
+export { default as ThemeIcon } from './ThemeIcon';
+export { default as ThemeSwitch } from './ThemeSwitch';
