+++ packages/theme/src/styles/components/ThemeIcon.scss
